Separate RoleResolver service
