+++ src/ORMInterface.php
@@ -12,6 +12,7 @@ use Cycle\ORM\Service\IndexProviderInterface;
 use Cycle\ORM\Service\MapperProviderInterface;
 use Cycle\ORM\Service\RelationProviderInterface;
 use Cycle\ORM\Service\RepositoryProviderInterface;
+use Cycle\ORM\Service\RoleResolverInterface;
 use Cycle\ORM\Service\SourceProviderInterface;
 use Cycle\ORM\Transaction\CommandGeneratorInterface;
 
@@ -25,13 +26,9 @@ interface ORMInterface extends
     MapperProviderInterface,
     RepositoryProviderInterface,
     RelationProviderInterface,
+    RoleResolverInterface,
     IndexProviderInterface
 {
-    /**
-     * Automatically resolve role based on object name or instance.
-     */
-    public function resolveRole(string|object $entity): string;
-
     /**
      * Create new entity based on given role and input data. Method will attempt to re-use
      * already loaded entity.
