+++ src/Service/Implementation/EntityFactory.php
@@ -4,8 +4,6 @@ declare(strict_types=1);
 
 namespace Cycle\ORM\Service\Implementation;
 
-use Cycle\ORM\EntityProxyInterface;
-use Cycle\ORM\Exception\ORMException;
 use Cycle\ORM\Heap\HeapInterface;
 use Cycle\ORM\Heap\Node;
 use Cycle\ORM\Service\EntityFactoryInterface;
@@ -14,6 +12,7 @@ use Cycle\ORM\Service\MapperProviderInterface;
 use Cycle\ORM\Service\RelationProviderInterface;
 use Cycle\ORM\SchemaInterface;
 use Cycle\ORM\Select\LoaderInterface;
+use Cycle\ORM\Service\RoleResolverInterface;
 
 /**
  * @internal
@@ -26,6 +25,7 @@ final class EntityFactory implements EntityFactoryInterface
         private MapperProviderInterface $mapperProvider,
         private RelationProviderInterface $relationProvider,
         private IndexProviderInterface $indexProvider,
+        private RoleResolverInterface $roleResolver,
     ) {}
 
     public function make(
@@ -37,7 +37,7 @@ final class EntityFactory implements EntityFactoryInterface
         $role = $data[LoaderInterface::ROLE_KEY] ?? $role;
         unset($data[LoaderInterface::ROLE_KEY]);
         // Resolved role
-        $rRole = $this->resolveRole($role);
+        $rRole = $this->roleResolver->resolveRole($role);
         $relMap = $this->relationProvider->getRelationMap($rRole);
         $mapper = $this->mapperProvider->getMapper($rRole);
 
@@ -78,32 +78,5 @@ final class EntityFactory implements EntityFactoryInterface
         $this->heap->attach($e, $node, $this->indexProvider->getIndexes($rRole));
 
         return $mapper->hydrate($e, $relMap->init($this, $node, $castedData));
-    }
-
-    public function resolveRole(object|string $entity): string
-    {
-        if (\is_object($entity)) {
-            $node = $this->heap->get($entity);
-            if ($node !== null) {
-                return $node->getRole();
-            }
-
-            $class = $entity::class;
-            if (!$this->schema->defines($class)) {
-                $parentClass = \get_parent_class($entity);
-
-                if ($parentClass === false
-                    || !$entity instanceof EntityProxyInterface
-                    || !$this->schema->defines($parentClass)
-                ) {
-                    throw new ORMException("Unable to resolve role of `$class`.");
-                }
-                $class = $parentClass;
-            }
-
-            $entity = $class;
-        }
-
-        return $this->schema->resolveAlias($entity) ?? throw new ORMException("Unable to resolve role `$entity`.");
     }
 }
